Move `getElementProperties` into separate file
